Module ratelimit_meter::algorithms::gcra

The Generic Cell Rate Algorithm

Structs

GCRA
Implements the virtual scheduling description of the Generic Cell Rate Algorithm, attributed to ITU-T in recommendation I.371 Traffic control and congestion control in B-ISDN; from Wikipedia.

NotUntil
Returned in case of a negative rate-limiting decision. Indicates the earliest instant that a cell might get accepted again.

State
The GCRA's state about a single rate limiting history.