Trait ratelimit_meter::clock::Reference

pub trait Reference: Sized + Add<Duration, Output = Self> + PartialEq + Eq + Ord + Copy + Clone + Send + Sync + Debug {
    fn duration_since(&self, earlier: Self) -> Duration;

    fn saturating_sub(&self, duration: Duration) -> Self;
}

A measurement from a clock.

Required methods

fn duration_since(&self, earlier: Self) -> Duration

Determines the time that separates two measurements of a clock. Implementations of this must perform a saturating subtraction - if the earlier timestamp should be later, duration_since must return the zero duration.

fn saturating_sub(&self, duration: Duration) -> Self

Returns a reference point that lies at most duration in the past from the current reference. If an underflow should occur, returns the current reference.

Loading content...