Module ratelimit_meter::state

Data structures that keep rate-limiting state.

Re-exports

pub use self::direct::DirectRateLimiter;

pub use self::keyed::KeyedRateLimiter;

Modules

direct
An in-memory rate limiter that can make decisions for a single situation.

keyed
An in-memory rate limiter that can keep track of rates for multiple keys, e.g. per-customer or per-IC rates.